Duplex browser, configuration and dependencies refactoring:
Changes:
- the browser is no longer the Duplex brain, but simply shows and activates controller "configurations" via processes (active configurations)
- a configuration is a table that launches a controller and one or more apps it wants to run in this configuration. this replaces the old multi app configs in the browser with the app aliases.
- configurations are defined in the /Controllers folders now, adding new controllers should be possible without changing anything in the Duplex core, making it hopefully much easier to add new controllers
- apps are implementation details of the controllers now, only directly referenced and configured in the controller files, so we can hopefully start adding new apps or more easily refactor the current ones (split them into more "sub" apps to allow easier configuration)

Fixes:
- switching between Applications while another device process is running now properly clears the Virtual UI

Note: The old app level config dialogs and settings are gone. They way they used to be, they added a lot of dependencies to the browser and apps. Lets please solve this after the beta with a new approach, preferably using renoise.Documents and using the current configs as base for user modifiied custom configs. 

Changing the browser was a bitch, so I ended up in completely rewriting it. Please let me know if I have broken something.